added iteration 6

+++ backend/controllers/bookController.js
@@ -52,9 +52,9 @@ const getBookById = async (req, res) => {
 const deleteBook = async (req, res) => {
   const { id } = req.params;
   try {
-    // const user_id = req.user._id;
-    // const book = await Book.findByIdAndDelete({ _id: id, user_id: user_id });
-    const book = await Book.findByIdAndDelete({ _id: id,});
+    const user_id = req.user._id;
+    const book = await Book.findByIdAndDelete({ _id: id, user_id: user_id }); // User can only delete their own books
+    // const book = await Book.findByIdAndDelete({ _id: id,});
     if (!book) {
       return res.status(404).json({ message: 'Book not found' });
     }
@@ -69,10 +69,10 @@ const deleteBook = async (req, res) => {
 const updateBook = async (req, res) => {
   const { id } = req.params;
   try {
-    // const user_id = req.user._id;
+    const user_id = req.user._id;
     const book = await Book.findOneAndUpdate(
-      // { _id: id, user_id: user_id },
-      { _id: id,  },
+      { _id: id, user_id: user_id },  // User can only update their own books
+      //{ _id: id,  },
       { ...req.body },
       { new: true }
     );
